#2181 Логика отвязки отдельных целей и задач. Убрал возможность на фронте отвязывать отдельные цели и задачи во "внешнем" проекте.

+++ lib/L1_domain/entities/task_ext_actions.dart
@@ -31,7 +31,7 @@ extension TaskActionsExt on Task {
     if (canImport) {
       res.add(TaskActionType.import);
     }
-    if (hasLink) {
+    if (isProject && hasLink) {
       res.add(TaskActionType.go2source);
       res.add(TaskActionType.unlink);
       res.add(TaskActionType.unwatch);
